requirements.txt
langchain==0.3.21
langchain-community==0.3.20
flask==3.1.0
flask-cors==5.0.1
pymongo==4.11.3
faiss-cpu==1.10.0
sentence-transformers==4.0.1
huggingface-hub==0.29.3
requests==2.32.3
python-dotenv==1.1.0